Add Intel 80386 CPU core (x86-32) (#4)
* Add Intel 80386 CPU core (x86-32)

New self-contained remu::x86_32 core: 32-bit registers and addressing
(ModRM+SIB), real mode, protected mode (descriptor loads, privilege
checks, call/interrupt/trap gates, TSS stack switching, V86 entry/exit,
I/O permission bitmap) and paging (two-level walk, A/D bits, #PF with
CR2, software TLB). Full 386 integer instruction set including the
0F map (bit ops, SHLD/SHRD, MOVZX/MOVSX, SETcc, BSF/BSR, system
instructions and CR/DR/TR moves).

The real-mode instruction set is validated against all 1,758,700 cases
of the hardware-generated SingleStepTests 80386 suite (941 MOO files,
tests/harte_80386.rs parses the MOO chunked format directly). Passing
required implementing 386 behavior well beyond the manuals:

- BT/BTS/BTR/BTC route the operand through the rotator: OF gets the
  ROR top-two-bits rule; memory bit displacements wrap at the address
  size.
- SHLD/SHRD counts past a 16-bit operand refeed the source
  (dst:src:src); AF is always set; SHRD's OF tracks the last 1-bit
  step.
- BSF/BSR "undefined" flags follow the internal negate/align ops
  (derived empirically from the suite; matches upstream issue #4).
- IDIV skips the classic magnitude pre-check: a lossy W-bit remainder
  register lets some overflowing quotients (exactly -2^(W-1)) escape
  #DE with hardware-exact results; dividend = minimum always faults.
- The invalid-SIB encodings (index=100, scale>0) scale the base
  register, except when the base is the mod=00 disp32.
- POPAD really pops the ESP slot (its high word survives the SP
  fixup); PUSHA writes the frame bottom-up and keeps completed writes
  on a mid-frame fault; string/POPA faults keep per-step progress.
- o32 POP sreg performs a 16-bit read but releases 4 stack bytes.
- Byte shifts by CL: counts 16/24 pull CF from the replicated byte
  lane; RCL/RCR iterate the full masked count (flags track the true
  last step).
- DAS carries the borrow of the AL-6 step into CF; AAA/AAS follow the
  documented AX+/-0x106 model (unlike the 8086); AAM 0 leaves the
  divider's PF signature (parity of AL*0x81).
- 386EX rig details in the harness: config ports 22h/23h read
  7Fh/42h; two-operand IMUL flags masked per upstream PR #6; the 8
  self-modifying tests that require prefetch-queue modeling are
  skipped and documented.

Protected mode and paging are covered by unit tests (PM entry, ring
transitions with TSS stack switch, gate delivery, #PF/CR2/A-D bits).
Approximate documented 386 timings; ~65-125M emulated instr/s in the
new cpu386 benchmarks.


* Fix protected-mode, paging and V86 defects found in review

Ten defects, all in the privilege-transition paths the SingleStepTests
80386 suite cannot reach (it is real mode only). Each fix ships with a
regression test that fails against the previous commit.

Paging and fault delivery:
- translate() classified every access by CPL, so the 386's *implicit*
  supervisor accesses — descriptor-table/IDT/TSS reads, accessed and
  busy-bit writebacks, and the frame pushes onto an inner-privilege
  stack made while CS still holds the outer segment — were page-checked
  as user accesses. With the standard OS layout (kernel structures
  mapped supervisor-only) the first ring-3 -> ring-0 interrupt raised a
  spurious #PF and triple-faulted. Added a `supervisor_override` flag,
  `sys_read*`/`sys_write8` helpers for system structures, and explicit
  supervisor windows around the inner-stack frame pushes.
- deliver() escalated ANY fault during delivery to #DF. Implemented the
  SDM Table 6-5 benign/contributory classification, so e.g. a timer
  interrupt hitting a swapped-out kernel stack now takes a recoverable
  #PF. A fault while delivering #DF is a triple fault, as before.
- Delivery is now atomic: raise() rolls back every register side effect
  (CR2 excepted) when it faults partway, so a nested fault is delivered
  from the pre-delivery architectural state. This is what let
  v86_interrupt() leak a cleared VM flag — the fault path deliberately
  preserves EFLAGS — and deliver the nested fault down the non-V86 path
  with a stale V86 CS cache and a malformed frame.

Protected mode:
- IRET to an outer ring committed the popped EFLAGS image before
  load_outer_ss(), which can still fault; a bad user SS entered the
  kernel's fault handler with the user's flags (TF, IF) already live.
  The commit now happens after the last faultable step.
- table() truncated the cached LDT limit to 16 bits, spuriously #GP'ing
  every selector past index 0x1000 in a page-granular LDT.
- IRETD with VM set in the popped image at CPL > 0 raised #GP; hardware
  simply ignores the bit (it is not in the flag-write mask).
- An out-of-range SS selector sourced from the TSS raised #GP instead of
  the SDM-mandated #TS (read_descriptor now takes the fault ctor).
- tss_stack() limit checks were off by one, reading the inner SS one
  byte past the TSS limit instead of raising #TS.
- Error codes generated while delivering an external interrupt now set
  the EXT bit (vector*8+3, selector|1).

Instruction behavior:
- A 32-bit REP could run 2^32 iterations with NMI/INTR blocked inside a
  single step(). It now polls between iterations and rewinds EIP to the
  prefix so the handler's IRET resumes it, as on hardware.

The 1,758,700-case hardware suite still passes unchanged.


* Clear pre-existing clippy lints in the 8086 core for green CI

A newer stable clippy flags four lints in src/x86/ (unnecessary cast,
manual is_multiple_of ×2, collapsible let-chain) that postdate the 8086
core's merge, so CI's `cargo clippy --lib -- -D warnings` step is red on
main independent of the 386 work. These are mechanical, behavior-
preserving fixes so this PR lands with a green build.


---------

+++ src/x86/execute.rs
@@ -190,7 +190,7 @@ impl Cpu {
             0x98 => { self.regs.ax = self.regs.ax as u8 as i8 as i16 as u16; 2 }
             0x99 => { self.regs.dx = if self.regs.ax & 0x8000 != 0 { 0xFFFF } else { 0 }; 5 }
             0x9E => { // SAHF
-                let ah = (self.regs.ax >> 8) as u16;
+                let ah = self.regs.ax >> 8;
                 let keep = self.regs.flags.bits() & 0xFF00;
                 self.regs.flags = Flags::from_bits_truncate(keep | ah);
                 4
